requirements.txt
colored
bs4
requests
python-pptx
pillow